refactor reason generation
