+++ databases/catdat/scripts/shared.ts
@@ -55,3 +55,33 @@ export function get_conclusion_string(
 
 	return `${properties_dict[conclusion][conditional ? 'conditional' : 'relation']} ${conclusion}`
 }
+
+export function get_reason_string(
+	implication: NormalizedImplication,
+	properties_dict: Record<string, PropertyMeta>,
+) {
+	const assumption_string = get_assumption_string(implication, properties_dict)
+	const conclusion_string = get_conclusion_string(implication, properties_dict)
+
+	const ref = `by <a href="/category-implication/${implication.id}">this result</a>`
+	return `Since it ${assumption_string}, it ${conclusion_string} (${ref}).`
+}
+
+export function get_contradiction_string(
+	implication: NormalizedImplication,
+	properties_dict: Record<string, PropertyMeta>,
+	property: string,
+) {
+	const assumption_string = get_assumption_string(implication, properties_dict, true)
+	const conclusion_string = get_conclusion_string(implication, properties_dict, true)
+
+	const has_multiple_assumptions = implication.assumptions.size > 1
+
+	const ref = `by <a href="/category-implication/${implication.id}">this result</a>`
+
+	const contra = `Assume for contradiction that it ${properties_dict[property].relation} ${property}`
+
+	return has_multiple_assumptions
+		? `${contra}. Then it ${assumption_string}, so it ${conclusion_string} (${ref}) – contradiction.`
+		: `${contra}. Then it ${conclusion_string} (${ref}) – contradiction.`
+}
